+++ src/main/java/org/mapfish/print/map/readers/WMTSMapReader.java
@@ -116,12 +116,12 @@ public class WMTSMapReader extends TileableMapReader {
 
     protected void renderTiles(TileRenderer formatter, Transformer transformer, URI commonUri, ParallelMapTileLoader parallelMapTileLoader) throws IOException, URISyntaxException {
         if (matrixIds != null) {
-            float diff = Float.POSITIVE_INFINITY;
-            float targetResolution = transformer.getGeoW() / transformer.getStraightBitmapW();
+            double diff = Double.POSITIVE_INFINITY;
+            double targetResolution = transformer.getGeoW() / transformer.getStraightBitmapW();
             for (int i = 0 ; i < matrixIds.size() ; i++) {
                 PJsonObject matrixId = matrixIds.getJSONObject(i);
                 float resolution = matrixId.getFloat("resolution");
-                float delta = Math.abs(1 - resolution / targetResolution);
+                double delta = Math.abs(1 - resolution / targetResolution);
                 if (delta < diff) {
                     diff = delta;
                     matrix = matrixId;
@@ -143,7 +143,7 @@ public class WMTSMapReader extends TileableMapReader {
         super.renderTiles(formatter, transformer, commonUri, parallelMapTileLoader);
     }
 
-    protected URI getTileUri(URI commonUri, Transformer transformer, float minGeoX, float minGeoY, float maxGeoX, float maxGeoY, long w, long h) throws URISyntaxException, UnsupportedEncodingException {
+    protected URI getTileUri(URI commonUri, Transformer transformer, double minGeoX, double minGeoY, double maxGeoX, double maxGeoY, long w, long h) throws URISyntaxException, UnsupportedEncodingException {
         if (matrixIds != null) {
             PJsonArray topLeftCorner = matrix.getJSONArray("topLeftCorner");
             float factor = 1 / (matrix.getFloat("resolution") * w);
@@ -185,7 +185,7 @@ public class WMTSMapReader extends TileableMapReader {
             }
         }
         else {
-            float targetResolution = (maxGeoX - minGeoX) / w;
+            double targetResolution = (maxGeoX - minGeoX) / w;
             WMTSLayerInfo.ResolutionInfo resolution = tileCacheLayerInfo.getNearestResolution(targetResolution);
 
             int col = (int) Math.round(Math.floor(((maxGeoX + minGeoX)/2-tileOrigin.getFloat(0)) / (resolution.value * w)));
